Create showtime
Should create showtime and should grab the movie data from the ProvidedApi.

Reserve seats

Reserving the seat response will contain a GUID of the reservation, also the number of seats, the auditorium used and the movie that will be played.
It should not be possible to reserve the same seats two times in 10 minutes.
It shouldn't be possible to reserve an already sold seat.
All the seats, when doing a reservation, need to be contiguous.